refactor(naming): move infra helpers utils/->lib/, rename route group, fix typo
- utils/axios-helper.ts -> lib/http.ts (api() + urls; name no longer misleading)
- utils/number-helper.ts -> lib/format.ts
- utils/datetime-helper.ts -> lib/date.ts
- utils/content-helper.ts -> lib/text.ts
- update all ~25 importers (alias paths) incl. the relative one in forestChange.store
- route group (group-ContainerWidth) -> (content); fix the 2 PageParams import strings
- fix type typo InvesmentTrackerParams -> InvestmentTrackerParams

Pure rename/move, no behavior change. tsc clean (after .next cache clear), build 31/31.
Domain helpers (country-helper, slug-iso2) + stores stay for PR3/PR6.
