+++ src/app/(content)/investment-tracker/[country]/richDataExample.json
