+++ src/app/(content)/investment-tracker/page.tsx
