+++ src/app/(content)/news/page.tsx
