+++ src/app/(content)/press/page.tsx
