+++ src/components/sections/features/forest-cover/AnnualPayout.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { PageParams } from "@/app/(group-ContainerWidth)/[country]/[year]/page";
+import { PageParams } from "@/app/(content)/[country]/[year]/page";
 import Br from "@/components/ui/Br";
 import { getCountryDetails } from "@/utils/country-helper";
 import { useParams } from "next/navigation";
