+++ src/components/sections/features/investment/InvestmentTrackerContent.tsx
@@ -1,9 +1,9 @@
 import Br from "@/components/ui/Br";
 import ContentSection from "@/components/ui/ContentSection";
-import { formatDateAgo } from "@/utils/datetime-helper";
+import { formatDateAgo } from "@/lib/date";
 import { InvestmentTrackerForCountry } from "@/utils/types";
 import RichToHTML from "./RichToHTML";
-import { hasContent } from "@/utils/content-helper";
+import { hasContent } from "@/lib/text";
 
 type Props = Partial<InvestmentTrackerForCountry> & {
   how_an_investment_could_work?: string;
