+++ src/lib/text.ts
